chinese-clip/requirements.txt
transformers>=4.18.0
torch>=1.9.0
numpy>=1.19.5
pillow>=8.0.0
fastapi>=0.88.0
uvicorn>=0.15.0
pydantic>=1.8.2
python-multipart>=0.0.5
requests>=2.25.1
